pyproject.toml
[tool.poetry]
name = "programmerassistantbot"
version = "0.1.0"
description = ""
authors = ["melloru <pythonubuntu12@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
aiogram = "^3.15.0"
python-dotenv = "^1.0.1"
sqlalchemy = "^2.0.36"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
